Rename legacyRunner to query-runner (kebab-case)
