+++ extensions/ql-vscode/src/data-extensions-editor/generate-flow-model.ts
@@ -1,7 +1,7 @@
 import { CancellationToken } from "vscode";
 import { DatabaseItem } from "../local-databases";
 import { join } from "path";
-import { QueryRunner } from "../query-server/queryRunner";
+import { QueryRunner } from "../query-server/query-runner";
 import { CodeQLCliServer } from "../cli";
 import { TeeLogger } from "../common";
 import { extensiblePredicateDefinitions } from "./predicates";
