+++ extensions/ql-vscode/src/debugger/debugger-factory.ts
@@ -11,7 +11,7 @@ import {
 import { isCanary } from "../config";
 import { LocalQueries } from "../local-queries";
 import { DisposableObject } from "../pure/disposable-object";
-import { QueryRunner } from "../query-server/queryRunner";
+import { QueryRunner } from "../query-server/query-runner";
 import { QLDebugConfigurationProvider } from "./debug-configuration";
 import { QLDebugSession } from "./debug-session";
