+++ extensions/ql-vscode/src/query-editor.ts
@@ -1,6 +1,6 @@
 import { Uri, window } from "vscode";
 import { CodeQLCliServer } from "./cli";
-import { QueryRunner } from "./query-server/queryRunner";
+import { QueryRunner } from "./query-server/query-runner";
 import { basename, join } from "path";
 import { getErrorMessage } from "./pure/helpers-pure";
 import { redactableError } from "./pure/errors";
