+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -49,7 +49,7 @@ import { EvalLogViewer } from "../eval-log-viewer";
 import EvalLogTreeBuilder from "../eval-log-tree-builder";
 import { EvalLogData, parseViewerData } from "../pure/log-summary-parser";
 import { QueryWithResults } from "../run-queries-shared";
-import { QueryRunner } from "../query-server/queryRunner";
+import { QueryRunner } from "../query-server/query-runner";
 import { VariantAnalysisManager } from "../variant-analysis/variant-analysis-manager";
 import { VariantAnalysisHistoryItem } from "./variant-analysis-history-item";
 import { getTotalResultCount } from "../variant-analysis/shared/variant-analysis";
