+++ extensions/ql-vscode/src/query-server/legacy/legacy-query-runner.ts
@@ -8,7 +8,11 @@ import {
   deregisterDatabases,
   registerDatabases,
 } from "../../pure/legacy-messages";
-import { CoreQueryResults, CoreQueryTarget, QueryRunner } from "../queryRunner";
+import {
+  CoreQueryResults,
+  CoreQueryTarget,
+  QueryRunner,
+} from "../query-runner";
 import { QueryOutputDir } from "../../run-queries-shared";
 import { QueryServerClient } from "./query-server-client";
 import {
