+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -24,7 +24,7 @@ import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { compileDatabaseUpgradeSequence } from "./upgrades";
 import { QueryEvaluationInfo, QueryOutputDir } from "../../run-queries-shared";
 import { redactableError } from "../../pure/errors";
-import { CoreQueryResults, CoreQueryTarget } from "../queryRunner";
+import { CoreQueryResults, CoreQueryTarget } from "../query-runner";
 import { Position } from "../../pure/messages-shared";
 
 export async function compileQuery(
