+++ extensions/ql-vscode/test/vscode-tests/cli-integration/queries.test.ts
@@ -20,7 +20,7 @@ import { describeWithCodeQL } from "../cli";
 import {
   CoreCompletedQuery,
   QueryRunner,
-} from "../../../src/query-server/queryRunner";
+} from "../../../src/query-server/query-runner";
 import { SELECT_QUERY_NAME } from "../../../src/contextual/locationFinder";
 import { LocalQueries } from "../../../src/local-queries";
 import { QueryResultType } from "../../../src/pure/new-messages";
